refactor: rename profile menu buttons
